+++ internal/provider/model_sdwan_service_switchport_feature.go
@@ -246,7 +246,7 @@ func (data ServiceSwitchport) toBody(ctx context.Context) string {
 			if item.EnableDot1x.IsNull() {
 				if true {
 					itemBody, _ = sjson.Set(itemBody, "enableDot1x.optionType", "default")
-					itemBody, _ = sjson.Set(itemBody, "enableDot1x.value", false)
+					itemBody, _ = sjson.Set(itemBody, "enableDot1x.value", true)
 				}
 			} else {
 				if true {