+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Bump `sdwan_service_lan_vpn_feature` schema to `20.18.0`
 - Bump `sdwan_service_lan_vpn_interface_svi_feature` schema to `20.18.0`
 - Bump `sdwan_service_routing_ospf_feature` schema to `20.18.0`
-- Bump `sdwan_service_lan_vpn_interface_svi_feature` schema to `20.18.0`
+- Bump `sdwan_service_route_policy_feature` schema to `20.18.0`
 - Bump `sdwan_service_switchport_feature` schema to `20.18.0`
 - Bump `sdwan_service_wireless_lan_feature` schema to `20.18.0`